refactor(settings): extract platform settings model
